tighten up API
